pyproject.toml
[project]
name = "elginvault-250331"
version = "0.1.0"
description = ""
authors = [
    {name = "Ryan Yang",email = "ryany1819@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"

